+++ qualibrate_runner/core/models/workflow.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class WorkflowStatus(BaseModel):
+    active: bool
+    nodes_completed: int
+    run_duration: float
